requirements.txt
ModernGL
ModernGL.ext.obj
Pillow
Pyrr
